+++ R/permChacko.R
@@ -1,14 +1,11 @@
 #' @title The Chacko test for order-restriction with permutation test
 #' @param x vector of numeric values
 #' @param n_perm number of permutations to calculate the p-value numerically
-#' @param verbose if \code{TRUE}, prints intermediate messages and output
-#' @return A named vector with the following elements:
-#' \describe{
-#'  \item{chisq_bar}{the test statistic}
-#'  \item{analytic_p-value}{the p-value calculated analytically}
-#'  \item{numeric_p-value}{the p-value calculated numerically}
-#'  \item{tabular_p-value}{the tabular p-value from Chacko (1963)}
-#' }
+#' @param verbosity if \code{TRUE}, prints intermediate messages and output
+#' @return A list containing the test statistic, p-values (analytic, numeric and
+#' tabular, when available), the number of permutations performed, the original
+#' data and the reduced data. Use [names()] and [str()] on the output for more
+#' details.
 #'
 #' @references
 #' Chacko, V. J. (1963). Testing homogeneity against ordered alternatives. The
@@ -18,20 +15,25 @@
 #' Sankhyā: The Indian Journal of Statistics, Series B, 185-190.
 #' @importFrom stats weighted.mean pchisq
 #' @examples
-#' ruxton22_12_07 <- c(6, 8, 4, 7, 3)
-#' chacko66_3 <- c(10L, 16L, 14L, 12L, 18L)
-#' chacko66_5 <- c(12L, 14L, 18L, 16L, 22L, 20L, 18L, 24L, 26L, 30L)
+#' ruxton221207 <- c(6, 8, 4, 7, 3)
+#' chacko66_sec3 <- c(10L, 16L, 14L, 12L, 18L)
+#' chacko66_sec5 <- c(12L, 14L, 18L, 16L, 22L, 20L, 18L, 24L, 26L, 30L)
 #'
-#' permChacko(ruxton22_12_07)
-#' permChacko(chacko66_3)
-#' permChacko(chacko66_5)
+#' permChacko(ruxton221207)
+#' permChacko(chacko66_sec3)
+#' permChacko(chacko66_sec5)
 #' @export
-permChacko <- function(x, n_perm = 1000L, verbose = FALSE) {
+permChacko <- function(x, n_perm = 1000L, verbosity = 0) {
+  if (!is.null(dim(x))) stop("Input must be a vector")
+  if (verbosity >= 1L) message("Reducing original vector")
   # Ordering and reducing vector
-  x_t <- reduceVector(x, verbose)
+  x_t <- reduceVector(x, verbosity)[["x_t"]]
   k <- length(x)
   chisq_bar <- chackoStatistic(x_t, n = sum(x), k)
 
+  if (verbosity >= 1L && n_perm > 0L) {
+    message("\nReducing ", n_perm, " permutations of original vector")
+  }
   # Calculating the mean of the current permutation
   perm_chisq_bar <- vapply(
     X = seq_len(n_perm),
@@ -40,7 +42,7 @@ permChacko <- function(x, n_perm = 1000L, verbose = FALSE) {
 
       # For each such permutation we can go through the ordering procedure and
       # calculate the test statistic according to equation 5.
-      perm_x_t <- reduceVector(perm_x, verbose)
+      perm_x_t <- reduceVector(perm_x, 0L)[["x_t"]]
       perm_chisq_bar <- chackoStatistic(perm_x_t, n = sum(perm_x), k)
       return(perm_chisq_bar)
     },
@@ -64,12 +66,19 @@ permChacko <- function(x, n_perm = 1000L, verbose = FALSE) {
   } else {
     table_p_value <- NA
   }
-  return(
-    c(
-      "chisq_bar" = chisq_bar,
-      "analytic_p-value" = anal_p_value,
-      "numeric_p-value" = perm_p_value,
-      "tabular_p-value" = table_p_value
-    )
+  if (verbosity >= 1L) message("\nTest statistics")
+  p_values <- c(
+    "analytic" = anal_p_value,
+    "numeric" = perm_p_value,
+    "tabular" = table_p_value
   )
+  out <- list(
+    statistic = chisq_bar,
+    p_values = p_values,
+    n_perm = n_perm,
+    observed_data = x,
+    reduced_data = x_t
+  )
+  class(out) <- "chacko_test"
+  return(out)
 }