+++ R/permutateX.R
@@ -1,6 +1,6 @@
 permutateX <- function(X) {
   N <- sum(X)
   K <- length(X)
-  perm_x <- unname(table(sample(K, size = N, replace = TRUE)))
+  perm_x <- as.vector(unname(table(sample(K, size = N, replace = TRUE))))
   return(perm_x)
 }