+++ NEWS.md
@@ -1,3 +1,13 @@
+# permChacko 0.2.0
+
+* Replaced `verbose` with `verbosity` in functions to allow for more control over the output ([issue #3](https://github.com/ocbe-uio/permChacko/issues/3)).
+* Exported `reduceVector()` ([issue #4](https://github.com/ocbe-uio/permChacko/issues/4)).
+* Added package documentation ([issue #5](https://github.com/ocbe-uio/permChacko/issues/5)).
+* Standardized dataset names
+* Created print methods for `permChacko()` and `reduceVector()` ([issue #7](https://github.com/ocbe-uio/permChacko/issues/7)).
+* Created summary method for `reduceVector()` ([issue #8](https://github.com/ocbe-uio/permChacko/issues/8)).
+* Added vignette ([issue #2](https://github.com/ocbe-uio/permChacko/issues/2)).
+
 # permChacko 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.